requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-WTF==0.14.2
googlemaps==3.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pymediawiki==0.6.1
pytest==4.3.0
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
